+++ app/src/main/java/com/hyeeyoung/wishboard/view/folder/screens/FolderFragment.kt
@@ -1,20 +1,20 @@
-package com.hyeeyoung.wishboard.view.noti
+package com.hyeeyoung.wishboard.view.folder.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.hyeeyoung.wishboard.databinding.FragmentNotiBinding
+import com.hyeeyoung.wishboard.databinding.FragmentFolderBinding
 
-class NotiFragment : Fragment() {
-    private lateinit var binding: FragmentNotiBinding
+class FolderFragment : Fragment() {
+    private lateinit var binding: FragmentFolderBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentNotiBinding.inflate(inflater, container, false)
+        binding = FragmentFolderBinding.inflate(inflater, container, false)
 
         return binding.root
     }
@@ -22,7 +22,7 @@ class NotiFragment : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            NotiFragment().apply {
+            FolderFragment().apply {
 
             }
     }
