+++ app/src/main/java/com/hyeeyoung/wishboard/view/noti/screens/NotiFragment.kt
@@ -1,20 +1,20 @@
-package com.hyeeyoung.wishboard.view.wish
+package com.hyeeyoung.wishboard.view.noti.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.hyeeyoung.wishboard.databinding.FragmentWishBinding
+import com.hyeeyoung.wishboard.databinding.FragmentNotiBinding
 
-class WishFragment : Fragment() {
-    private lateinit var binding: FragmentWishBinding
+class NotiFragment : Fragment() {
+    private lateinit var binding: FragmentNotiBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentWishBinding.inflate(inflater, container, false)
+        binding = FragmentNotiBinding.inflate(inflater, container, false)
 
         return binding.root
     }
@@ -22,7 +22,7 @@ class WishFragment : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            WishFragment().apply {
+            NotiFragment().apply {
 
             }
     }
