+++ app/src/main/java/com/hyeeyoung/wishboard/view/sign/screens/SignUpEmailFragment.kt
@@ -1,4 +1,4 @@
-package com.hyeeyoung.wishboard.view.sign
+package com.hyeeyoung.wishboard.view.sign.screens
 
 import android.os.Bundle
 import android.view.LayoutInflater
