+++ app/src/main/java/com/hyeeyoung/wishboard/view/wish/item/screens/WishFragment.kt
@@ -1,20 +1,20 @@
-package com.hyeeyoung.wishboard.view.folder
+package com.hyeeyoung.wishboard.view.wish.item.screens
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.hyeeyoung.wishboard.databinding.FragmentFolderBinding
+import com.hyeeyoung.wishboard.databinding.FragmentWishBinding
 
-class FolderFragment : Fragment() {
-    private lateinit var binding: FragmentFolderBinding
+class WishFragment : Fragment() {
+    private lateinit var binding: FragmentWishBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentFolderBinding.inflate(inflater, container, false)
+        binding = FragmentWishBinding.inflate(inflater, container, false)
 
         return binding.root
     }
@@ -22,7 +22,7 @@ class FolderFragment : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FolderFragment().apply {
+            WishFragment().apply {
 
             }
     }
