+++ app/src/main/java/com/hyeeyoung/wishboard/view/wish/item/screens/WishItemDetailFragment.kt
@@ -1,9 +1,8 @@
-package com.hyeeyoung.wishboard.view.home.wishitem
+package com.hyeeyoung.wishboard.view.wish.item.screens
 
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
