+++ app/src/main/res/layout/fragment_cart.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.cart.CartFragment">
+    tools:context=".view.cart.screens.CartFragment">
 
     <data>
