+++ app/src/main/res/layout/fragment_my.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.my.MyFragment">
+    tools:context=".view.my.screens.MyFragment">
 
     <data>
