+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.sign.SignInFragment">
+    tools:context=".view.sign.screens.SignInFragment">
 
     <data>
