+++ app/src/main/res/navigation/nav_graph.xml
@@ -8,7 +8,7 @@
     <include app:graph="@navigation/wish_item_nav_graph" />
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.hyeeyoung.wishboard.view.home.HomeFragment"
+        android:name="com.hyeeyoung.wishboard.view.wish.list.screens.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home">
         <argument
@@ -24,7 +24,7 @@
     </fragment>
     <fragment
         android:id="@+id/cartFragment"
-        android:name="com.hyeeyoung.wishboard.view.cart.CartFragment"
+        android:name="com.hyeeyoung.wishboard.view.cart.screens.CartFragment"
         android:label="fragment_sign_up_email"
         tools:layout="@layout/fragment_sign_up_email">
         <action
@@ -33,17 +33,17 @@
     </fragment>
     <fragment
         android:id="@+id/wishFragment"
-        android:name="com.hyeeyoung.wishboard.view.wish.WishFragment"
+        android:name="com.hyeeyoung.wishboard.view.wish.item.screens.WishFragment"
         android:label="fragment_wish"
         tools:layout="@layout/fragment_wish" />
     <fragment
         android:id="@+id/folderFragment"
-        android:name="com.hyeeyoung.wishboard.view.folder.FolderFragment"
+        android:name="com.hyeeyoung.wishboard.view.folder.screens.FolderFragment"
         android:label="fragment_folder"
         tools:layout="@layout/fragment_folder" />
     <fragment
         android:id="@+id/notiFragment"
-        android:name="com.hyeeyoung.wishboard.view.noti.NotiFragment"
+        android:name="com.hyeeyoung.wishboard.view.noti.screens.NotiFragment"
         android:label="fragment_noti"
         tools:layout="@layout/fragment_noti">
         <action
@@ -52,7 +52,7 @@
     </fragment>
     <fragment
         android:id="@+id/myFragment"
-        android:name="com.hyeeyoung.wishboard.view.my.MyFragment"
+        android:name="com.hyeeyoung.wishboard.view.my.screens.MyFragment"
         android:label="fragment_my"
         tools:layout="@layout/fragment_my" />
 </navigation>
